+++ webtest/forms.py
@@ -433,10 +433,15 @@ class Form:
         field_order = []
         tags = ('input', 'select', 'textarea', 'button')
         inner_elts = self.html.find_all(tags)
-        def _form_elt_filter(tag):
-            return tag in inner_elts or (
-                tag.attrs.get('form') == self.id and tag.name in tags)
-        elements = self.response.html.find_all(_form_elt_filter)
+        if self.id:
+            def _form_elt_filter(tag):
+                return tag.name in tags and any(
+                    prt.name == 'form' and prt.attrs.get('id') == self.id
+                    for prt in tag.parents
+                ) or tag.attrs.get('form') == self.id
+            elements = self.response.html.find_all(_form_elt_filter)
+        else:
+            elements = inner_elts
         for pos, node in enumerate(elements):
             attrs = dict(node.attrs)
             tag = node.name